types: Use PETScStruct which inherits from CCompositeObject
