+++ devito/ir/iet/nodes.py
@@ -1052,6 +1052,8 @@ class Dereference(ExprStmt, Node):
 
     @property
     def expr_symbols(self):
+        # TODO: Erick made a similar edit for NoDeclStruct so once merged,
+        # adjust accordingly
         if self.pointer.is_Object:
             ret = [self.pointer._C_symbol]
         else:
