+++ devito/petsc/iet/utils.py
@@ -1,5 +1,4 @@
 from devito.ir.iet.nodes import Call
-from devito.types import PETScStruct
 
 
 def petsc_call(specific_call, call_args):
@@ -13,7 +12,7 @@ def petsc_call_mpi(specific_call, call_args):
 
 
 def petsc_struct(name, fields, liveness='lazy'):
-    # pfields = [(i._C_name, i._C_ctype) for i in fields]
-    # from IPython import embed; embed()
+    # TODO: Fix this circular import
+    from devito.petsc.types.object import PETScStruct
     return PETScStruct(name=name, pname='MatContext',
-                            fields=fields, liveness=liveness)
+                       fields=fields, liveness=liveness)
