+++ devito/petsc/types/__init__.py
@@ -1,3 +1,3 @@
-from .types import * # noqa
-from .array import * # noqa
-from .object import * # noqa
+from .array import *  # noqa
+from .types import *  # noqa
+from .object import *  # noqa
