+++ devito/petsc/types/object.py
@@ -1,3 +1,5 @@
+from ctypes import POINTER
+
 from devito.tools import CustomDtype
 from devito.types import LocalObject, CCompositeObject
 from devito.symbolics import Byref
@@ -119,17 +121,22 @@ class DummyArg(LocalObject):
     dtype = CustomDtype('void', modifier='*')
 
 
-# class PETScStruct(CCompositeObject):
+class PETScStruct(CCompositeObject):
 
-#     __rargs__ = ('name', 'pname', 'fields')
+    __rargs__ = ('name', 'pname', 'fields')
 
-#     def __init__(self, name, pname, fields, liveness='lazy'):
-#         pfields = [(i._C_name, i._C_ctype) for i in fields]
-#         super().__init__(name, pname, pfields, liveness)
-#         self._fields = fields
+    def __init__(self, name, pname, fields, liveness='lazy'):
+        pfields = [(i._C_name, i._C_ctype) for i in fields]
+        super().__init__(name, pname, pfields, liveness)
+        self._fields = fields
 
-#     @property
-#     def fields(self):
-#         return self._fields
+    @property
+    def fields(self):
+        return self._fields
 
+    @property
+    def _C_ctype(self):
+        return POINTER(self.dtype) if self.liveness == \
+            'eager' else self.dtype
 
+    _C_modifier = ' *'
